Rename Features::StrictInitialize to Features::Initialize::Strict
