+++ test/micro/attributes/features_test.rb
@@ -57,55 +57,55 @@ class Micro::Attributes::FeaturesTest < Minitest::Test
     assert_includes(A.ancestors, ::Micro::Attributes)
     assert_includes(A.ancestors, Features::Diff)
     refute_includes(A.ancestors, Features::Initialize)
-    refute_includes(C.ancestors, Features::StrictInitialize)
+    refute_includes(C.ancestors, Features::Initialize::Strict)
     refute_includes(A.ancestors, Features::ActiveModelValidations)
 
     assert_includes(B.ancestors, ::Micro::Attributes)
     assert_includes(B.ancestors, Features::Initialize)
     refute_includes(B.ancestors, Features::Diff)
-    refute_includes(C.ancestors, Features::StrictInitialize)
+    refute_includes(C.ancestors, Features::Initialize::Strict)
     refute_includes(B.ancestors, Features::ActiveModelValidations)
 
     assert_includes(C.ancestors, ::Micro::Attributes)
     assert_includes(C.ancestors, Features::Initialize)
     assert_includes(C.ancestors, Features::Diff)
-    refute_includes(C.ancestors, Features::StrictInitialize)
+    refute_includes(C.ancestors, Features::Initialize::Strict)
     refute_includes(C.ancestors, Features::ActiveModelValidations)
 
     assert_includes(CStrict.ancestors, ::Micro::Attributes)
     assert_includes(CStrict.ancestors, Features::Initialize)
     assert_includes(CStrict.ancestors, Features::Diff)
-    assert_includes(CStrict.ancestors, Features::StrictInitialize)
+    assert_includes(CStrict.ancestors, Features::Initialize::Strict)
     refute_includes(CStrict.ancestors, Features::ActiveModelValidations)
 
     assert_includes(D.ancestors, ::Micro::Attributes)
     refute_includes(D.ancestors, Features::Diff)
     assert_includes(D.ancestors, Features::Initialize)
-    refute_includes(D.ancestors, Features::StrictInitialize)
+    refute_includes(D.ancestors, Features::Initialize::Strict)
     assert_includes(D.ancestors, Features::ActiveModelValidations)
 
     assert_includes(DStrict.ancestors, ::Micro::Attributes)
     assert_includes(DStrict.ancestors, Features::Initialize)
     refute_includes(DStrict.ancestors, Features::Diff)
-    assert_includes(DStrict.ancestors, Features::StrictInitialize)
+    assert_includes(DStrict.ancestors, Features::Initialize::Strict)
     assert_includes(DStrict.ancestors, Features::ActiveModelValidations)
 
     assert_includes(E.ancestors, ::Micro::Attributes)
     refute_includes(E.ancestors, Features::Initialize)
     assert_includes(E.ancestors, Features::Diff)
-    refute_includes(E.ancestors, Features::StrictInitialize)
+    refute_includes(E.ancestors, Features::Initialize::Strict)
     assert_includes(E.ancestors, Features::ActiveModelValidations)
 
     assert_includes(F.ancestors, ::Micro::Attributes)
     assert_includes(F.ancestors, Features::Diff)
     assert_includes(F.ancestors, Features::Initialize)
-    refute_includes(F.ancestors, Features::StrictInitialize)
+    refute_includes(F.ancestors, Features::Initialize::Strict)
     assert_includes(F.ancestors, Features::ActiveModelValidations)
 
     assert_includes(FStrict.ancestors, ::Micro::Attributes)
     assert_includes(FStrict.ancestors, Features::Initialize)
     assert_includes(FStrict.ancestors, Features::Diff)
-    assert_includes(FStrict.ancestors, Features::StrictInitialize)
+    assert_includes(FStrict.ancestors, Features::Initialize::Strict)
     assert_includes(FStrict.ancestors, Features::ActiveModelValidations)
   end
